+++ vagrant/install-ubuntu-20.04.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -e
+
 # Script to customize an Ubuntu 20.04 distribution so it's ready for a WoD usage
 # This first part is distribution specific and should be adapted based on its nature
 
@@ -7,5 +9,8 @@
 apt install -y ansible git openssh-server jq
 
 # Create the jupyter user
-userdel -f -r jupyter
+grep -qE '^jupyter:' /etc/passwd
+if [ $? -eq 0 ]; then
+        userdel -f -r jupyter
+fi
 useradd -U -m jupyter
